+++ main/src/desafios/exercicios_Iniciante/06-dobroTriploRaizQuadrada.md
