+++ main/src/desafios/exercicios_Iniciante/08-conversorMedidas.md
@@ -1,0 +1,11 @@
+
+
+
+
+
+
+### 08 — Conversor de Medidas
+
+**Exercício Python 008**: Escreva um programa que leia um valor em metros, e o exiba convertido em centímetros e milímetros.
+
+
