+++ main/src/desafios/exercicios_Iniciante/09-Tabuada.md
@@ -1,0 +1,11 @@
+### 
+
+### 12 — Calculando Descontos
+
+
+
+### 12 — Calculando Descontos
+
+**Exercício 012**: faça um algoritmo que leia o preço de um produto e mostre o seu novo preço, com 5% de desconto.
+
+
