+++ main/src/desafios/exercicios_Iniciante/12-calculandoDescontos.md
@@ -1,0 +1,11 @@
+
+
+
+
+### 13 — Reajuste Salarial
+
+**Exercício 013**: Faça um algoritmo que leia o salário de um funcionário e   mostre o seu novo salário, com 15% de aumento
+
+
+
+
