+++ main/src/desafios/exercicios_Iniciante/13-reajusteSalarial.md
@@ -1,0 +1,9 @@
+
+
+
+
+### 15 – Aluguel de Carros
+
+**Exercício 015**:   Escreva um programa que pergunte a quantidade de Km percorridos por um carro alugado e a quantidade de dias pelos   quais ele foi alugado. Calcule o preço a pagar, sabendo que o carro custa R$60 por dia e R60pordiaeR60 por dia e R60pordiaeR$0,15 por Km rodado.
+
+
