+++ main/src/desafios/resolvidos_Iniciante/06-dobroTriploRaizQuadrada.kt
@@ -1,4 +1,4 @@
-package desafios.iniciante
+package desafios
 
 import kotlin.math.sqrt
