+++ main/src/desafios/resolvidos_Iniciante/13-reajusteSalarial.kt
@@ -1,4 +1,4 @@
-package desafios.iniciante
+package desafios.resolvidos_Iniciante
 
 /*
 13 — Reajuste Salarial
